Authorization of product actions completed

+++ back-end/src/controllers/Urun.js
@@ -15,12 +15,12 @@ const kullaniciFavoriEkle = function (req, res) {
         UrunSema.findById(urunid)
             .then(urun => {
                 if (!urun) {
-                    cevapOlustur(res, 404, { "durum": "urun bulunamadi" });
+                    cevapOlustur(res, 404, { "hata": "urun bulunamadi" });
                 } else {
                     KullaniciSema.findById(userid).select("favoriler")
                         .then(favoriler => {
                             if (favoriler.favoriler.some(fav => fav.equals(urunid))) {
-                                cevapOlustur(res, 400, { "durum": "urun zaten favorilerde bulunuyor" });
+                                cevapOlustur(res, 400, { "hata": "urun zaten favorilerde bulunuyor" });
                             } else {
                                 favoriler.favoriler.push(urun);
                                 favoriler.save()
@@ -28,12 +28,12 @@ const kullaniciFavoriEkle = function (req, res) {
                                         cevapOlustur(res, 201, response);
                                     })
                                     .catch(err => {
-                                        cevapOlustur(res, 400, { "durum": "eklenirken bir sorun olustu" });
+                                        cevapOlustur(res, 400, { "hata": "eklenirken bir sorun olustu" });
                                     })
                             }
                         })
                         .catch(err => {
-                            cevapOlustur(res, 404, { "durum": "kullanici bulunamadi" });
+                            cevapOlustur(res, 404, { "hata": "kullanici bulunamadi" });
                         })
                 }
             })
@@ -54,7 +54,7 @@ const kullaniciFavorileriGetir = function (req, res) {
                 if (favoriler) {
                     cevapOlustur(res, 200, favoriler);
                 } else {
-                    cevapOlustur(res, 404, { "durum": "kullanici bulunamadi" });
+                    cevapOlustur(res, 404, { "hata": "kullanici bulunamadi" });
                 }
             })
             .catch(err => cevapOlustur(res, 400, err))
@@ -72,18 +72,18 @@ const kullaniciFavoriSil = function (req, res) {
             .then(async favoriliste => {
                 if (favoriliste) {
                     if (!favoriliste.favoriler.some(fav => fav.equals(urunid))) {
-                        cevapOlustur(res, 404, { "durum": "ilgili urun favorilerde yok" });
+                        cevapOlustur(res, 404, { "hata": "ilgili urun favorilerde yok" });
                     } else {
                         favoriliste.favoriler.pull(urunid)
                         try {
                             await favoriliste.save();
-                            cevapOlustur(res, 200, { "durum": `${urunid}'li urun favorilerden silindi` });
+                            cevapOlustur(res, 200, { "hata": `${urunid}'li urun favorilerden silindi` });
                         } catch (error) {
                             cevapOlustur(res, 400, error)
                         }
                     }
                 } else {
-                    cevapOlustur(res, 404, { "durum": "kullanici bulunamadi" })
+                    cevapOlustur(res, 404, { "hata": "kullanici bulunamadi" })
                 }
             })
             .catch(err => {
@@ -100,12 +100,12 @@ const urunAra = function (req, res) {
     UrunSema.find({ "urunAdi": { "$regex": urun, "$options": "i" } })  // i: case insensitive
         .then(urunler => {
             if (urunler.length == 0) {
-                cevapOlustur(res, 404, { "durum": "Aradiginz kelime ile eslesen urun bulunamadi" });
+                cevapOlustur(res, 404, { "hata": "Aradiginz kelime ile eslesen urun bulunamadi" });
             } else if (urunler) {
                 cevapOlustur(res, 200, urunler);
             }
             else {
-                cevapOlustur(res, 404, { "durum": "urun bulunamadi" });
+                cevapOlustur(res, 404, { "hata": "urun bulunamadi" });
             }
         })
         .catch(err => {
@@ -113,38 +113,43 @@ const urunAra = function (req, res) {
         })
 }
 const urunEkle = async function (req, res) {
-    const urunAdi = req.body.urunAdi;
-    const urunDetay = req.body.urunDetay;
-    const urunOzellikler = req.body.urunOzellikler ? req.body.urunOzellikler.split(",").map(item => item.trim()) : [];
-    const resimUrl = req.body.resimUrl;
-    const urunFiyat = req.body.urunFiyat;
-    const stok = req.body.stok;
-    const kategori = req.body.kategori;
+    if (req.auth.otorite == "admin") {
+        const urunAdi = req.body.urunAdi;
+        const urunDetay = req.body.urunDetay;
+        const urunOzellikler = req.body.urunOzellikler ? req.body.urunOzellikler.split(",").map(item => item.trim()) : [];
+        const resimUrl = req.body.resimUrl;
+        const urunFiyat = req.body.urunFiyat;
+        const stok = req.body.stok;
+        const kategori = req.body.kategori;
 
-    if (!urunAdi || !resimUrl || !urunFiyat || !stok || !kategori) {
-        cevapOlustur(res, 400, { "hata": "Bütün alanlar gereklidir" })
-        return;
-    }
-    try {
-        const mevcutUrun = await UrunSema.findOne({ urunAdi: urunAdi });
-
-        if (mevcutUrun) {
-            cevapOlustur(res, 400, "Bu isimde bir ürün zaten bulunmakta!");
-        } else {
-            const urun = await UrunSema.create({
-                urunAdi: urunAdi,
-                urunDetay: urunDetay,
-                urunOzellikler: urunOzellikler,
-                resimUrl: resimUrl,
-                urunFiyat: urunFiyat,
-                stok: stok,
-                kategori: kategori
-            });
-            cevapOlustur(res, 201, urun);
+        if (!urunAdi || !resimUrl || !urunFiyat || !stok || !kategori) {
+            cevapOlustur(res, 400, { "hata": "Bütün alanlar gereklidir" })
+            return;
         }
-    } catch (error) {
-        cevapOlustur(res, 400, error);
+        try {
+            const mevcutUrun = await UrunSema.findOne({ urunAdi: urunAdi });
+
+            if (mevcutUrun) {
+                cevapOlustur(res, 400, {"hata":"Bu isimde bir ürün zaten bulunmakta!"});
+            } else {
+                const urun = await UrunSema.create({
+                    urunAdi: urunAdi,
+                    urunDetay: urunDetay,
+                    urunOzellikler: urunOzellikler,
+                    resimUrl: resimUrl,
+                    urunFiyat: urunFiyat,
+                    stok: stok,
+                    kategori: kategori
+                });
+                cevapOlustur(res, 201, urun);
+            }
+        } catch (error) {
+            cevapOlustur(res, 400, error);
+        }
+    } else {
+        cevapOlustur(res, 401, { "hata": "yetkiniz yok" });
     }
+
 }
 
 const urunGetir = async function (req, res) {
@@ -191,39 +196,44 @@ const urunDetayGetir = async function (req, res) {
 
 const urunGuncelle = async function (req, res) {
     const urunid = req.params.urunid;
-    const urunAdi = req.body.urunAdi;
-    const urunDetay = req.body.urunDetay;
-    const urunOzellikler = req.body.urunOzellikler ? req.body.urunOzellikler.split(",").map(item => item.trim()) : [];
-    const resimUrl = req.body.resimUrl;
-    const urunFiyat = req.body.urunFiyat;
-    const stok = req.body.stok;
-    const kategori = req.body.kategori;
+    if (req.auth.otorite == "admin") {
+        const urunAdi = req.body.urunAdi;
+        const urunDetay = req.body.urunDetay;
+        const urunOzellikler = req.body.urunOzellikler ? req.body.urunOzellikler.split(",").map(item => item.trim()) : [];
+        const resimUrl = req.body.resimUrl;
+        const urunFiyat = req.body.urunFiyat;
+        const stok = req.body.stok;
+        const kategori = req.body.kategori;
 
-    if (!urunid || !urunAdi || !resimUrl || !urunFiyat || !stok || !kategori) {
-        cevapOlustur(res, 400, { "durum": "butun alanlari doldur" });
-        return;
-    }
-    else {
-        try {
-            const urun = await UrunSema.findById(urunid);
-            urun.urunAdi = urunAdi;
-            urun.urunDetay = urunDetay;
-            urun.urunOzellikler = urunOzellikler;
-            urun.resimUrl = resimUrl;
-            urun.urunFiyat = urunFiyat;
-            urun.stok = stok;
-            urun.kategori = kategori;
-            try {
-                const save = await urun.save();
-                cevapOlustur(res, 200, save);
-            } catch (error) {
-                cevapOlustur(res, 400, error);
-            }
-
-        } catch (error) {
-            cevapOlustur(res, 400, error)
+        if (!urunid || !urunAdi || !resimUrl || !urunFiyat || !stok || !kategori) {
+            cevapOlustur(res, 400, { "hata": "butun alanlari doldur" });
+            return;
         }
+        else {
+            try {
+                const urun = await UrunSema.findById(urunid);
+                urun.urunAdi = urunAdi;
+                urun.urunDetay = urunDetay;
+                urun.urunOzellikler = urunOzellikler;
+                urun.resimUrl = resimUrl;
+                urun.urunFiyat = urunFiyat;
+                urun.stok = stok;
+                urun.kategori = kategori;
+                try {
+                    const save = await urun.save();
+                    cevapOlustur(res, 200, save);
+                } catch (error) {
+                    cevapOlustur(res, 400, error);
+                }
+
+            } catch (error) {
+                cevapOlustur(res, 400, error)
+            }
+        }
+    } else {
+        cevapOlustur(res, 401, { "hata": "yetkiniz yok" });
     }
+
 
 
 }
@@ -231,18 +241,22 @@ const urunGuncelle = async function (req, res) {
 
 const urunSil = async function (req, res) {
     const urunid = req.params.urunid;
+    if (req.auth.otorite == "admin") {
+        try {
+            const urun = await UrunSema.findByIdAndRemove(urunid);
+            if (!urun) {
+                return cevapOlustur(res, 404, { "hata": "urun bulunamadi" });
+            }
 
-
-    try {
-        const urun = await UrunSema.findByIdAndRemove(urunid);
-        if (!urun) {
-            return cevapOlustur(res, 404, { "durum": "urun bulunamadi" });
+            return cevapOlustur(res, 200, { "hata": "urun basariyla silindi" });
+        } catch (error) {
+            return cevapOlustur(res, 500, { "hata": "bir hata olustu" });
         }
-
-        return cevapOlustur(res, 200, { "durum": "urun basariyla silindi" });
-    } catch (error) {
-        return cevapOlustur(res, 500, { "durum": "bir hata olustu" });
+    } else {
+        cevapOlustur(res, 401, { "hata": "yetkiniz yok" });
     }
+
+
 }
 module.exports = {
     kullaniciFavoriEkle,
